+++ src/app/constants.js
@@ -1,3 +1,8 @@
+/**
+ * This file defines the current constant values.
+ * To deprecate a value, add it to ./constants-old.js
+ * with a corresponding log message.
+ */
 export const Component = {
   DESCRIPTION: 'description',
   STATUS: 'status',
@@ -12,5 +17,8 @@ export const FileType = {};
  * to help prevent typos.
  */
 export const CoordinationType = {
-  DATASET: 'dataset'
+  DATASET: 'dataset',
+  OBS_TYPE: 'obsType',
+  FEATURE_TYPE: 'featureType',
+  FEATURE_VALUE_TYPE: 'featureValueType',
 };